+++ problems/0209.ts
@@ -19,6 +19,8 @@ Output: 0
  
  */
 
+import {findGeaterEqualDescend} from "../sort/binarySearch";
+
 function minSubArrayLen(target: number, nums: number[]): number {
     const arrDesc = [];
     let base = 0;
@@ -27,10 +29,9 @@ function minSubArrayLen(target: number, nums: number[]): number {
         base += num;
         arrDesc.push(num - base);
         const targetNew = target - base;
-        const findIndex = find(arrDesc, targetNew);
+        const findIndex = findGeaterEqualDescend(arrDesc, targetNew);
         if (findIndex > -1) {
             const retNew = arrDesc.length - findIndex;
-            console.log(`val ${num} index ${index} findIndex: ${findIndex}, retNew: ${retNew} base: ${base}, targetNew: ${targetNew} arrDesc: ${arrDesc}`);
             if (ret === 0 || retNew < ret) {
                 ret = retNew;
             }
@@ -38,29 +39,5 @@ function minSubArrayLen(target: number, nums: number[]): number {
     });
     return ret;
 };
-
-/**
- * Find the index of value that is equal or greater than target in a sorted array
- * @param numsDescend the array of nums sorted in descending order
- * @param target the target value to find
- */
-function find(numsDescend: number[], target: number): number {
-    let left = -1;
-    let right = numsDescend.length - 1;
-
-    while (left < right) {
-        const mid = Math.ceil((left + right) / 2);
-        if (numsDescend[mid] < target) {
-            right = mid - 1;
-        } else {
-            left = mid;
-        }
-    }
-    if (right > -1) {
-        return numsDescend[right] >= target ? right : -1;
-    }
-
-    return -1;
-}
 
 console.log(minSubArrayLen(7, [2,3,1,2,4,3])); // expect 2
